:truck: Move CountdownContainer to `components` directory
